requirements-dev.txt
sanic
argon2
pyjwt
sqlalchemy
pillow
pyresttest
